Protecting video routes and fix deleteVideo

+++ controllers/videoController.js
@@ -72,7 +72,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render('editVideo', {pageTitle: `Edit ${video.title}`, video});
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      res.render('editVideo', {pageTitle: `Edit ${video.title}`, video});
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -97,7 +101,15 @@ export const deleteVideo = async (req, res) => {
     params: {id},
   } = req;
   try {
-    await Video.findOneAndRemove({_id: id});
+    const video = await Video.findById(id);
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      const index = req.user.videos.indexOf(id);
+      req.user.videos.splice(index, 1);
+      req.user.save();
+      await Video.findOneAndRemove({_id: id});
+    }
   } catch (error) {
     console.log(error);
   }
